+++ env.demo
@@ -43,3 +43,5 @@ WMTS_LAYER_ALTERNATIVE=ch.swisstopo.swissimage
 PRIVATE_DOCUMENTS_DIR=/var/sig/private_geocity/geocity_demo
 ALLOWED_HOSTS=localhost,127.0.0.1,form-demo.mapnv.ch,srvdev
 ALLOWED_CORS=http://localhost:3000,http://127.0.0.1
+# This setting will enable the factor authentification
+ENABLE_2FA=true
