+++ templates/base_generic.html
@@ -35,7 +35,9 @@
         <div class="dropdown-menu dropdown-menu-right" aria-labelledby="navbarDropdownMenuLink">
           <a class="dropdown-item" href="{% url 'password_change' %}">{% trans "Changer mon mot de passe" %}</a>
           <a class="dropdown-item" href="{% url 'permit_author_edit' %}">{% trans "Modifier mon compte" %}</a>
+          {% if ENABLE_2FA %}
           <a class="dropdown-item" href="{% url 'two_factor:profile' %}">{% trans "Account Security" %}</a>
+          {% endif %}
           <a class="dropdown-item" href="{% url 'logout' %}">{% trans "Me déconnecter" %}</a>
         </div>
       </div>
@@ -54,9 +56,11 @@
           <li class="nav-item">
             <a class="nav-link" href="{% url 'permit_author_edit' %}">{% trans "Modifier mon compte" %}</a>
           </li>
+          {% if ENABLE_2FA %}
           <li class="nav-item">
             <a class="nav-link" href="{% url 'two_factor:profile' %}">{% trans "Account Security" %}</a>
           </li>
+          {% endif %}
           <li class="nav-item">
             <a class="nav-link" href="{% url 'logout' %}">{% trans "Me déconnecter" %}</a>
           </li>
