+++ templates/two_factor/login.html
@@ -1,0 +1,81 @@
+<!-- This templates replaces the default django-two-factor-auth template at two_factor/core/login.html -->
+{% extends "base_generic.html" %}
+
+{% load i18n %}
+{% load i18n two_factor %}
+{% load bootstrap4 %}
+{% load static %}
+
+{% block stylesheets%}
+  <link rel="stylesheet" type="text/css" href="{% static 'css/login.css' %}">
+  <link rel="stylesheet" type="text/css" href="{% static 'css/permits.css' %}">
+{% endblock %}
+{% block content-fluid %}
+  <div id="login-welcome-text">
+    <h3 class="text-center">{{ config.APPLICATION_TITLE }}</h3>
+    <h5 class="text-center">{{ config.APPLICATION_SUBTITLE }}</h5>
+    <br>
+    <div class="text-center">{{ config.APPLICATION_DESCRIPTION }}</div>
+  </div>
+
+  {% if next and user.is_authenticated %}
+    <div class="text-center">
+      <p id="login-issue">{% trans "Votre compte ne dispose pas des droits nécessaires pour accéder à cette page" %}</p>
+    </div>
+  {% endif %}
+
+  {% if wizard.steps.current == 'auth' %}
+    {% if form.errors %}
+      <div class="text-center mt-3">
+        <p class="text-danger font-weight-bold">{% trans "Votre mot de passe et votre nom d'utilisateur ne correspondent pas" %}</p>
+      </div>   
+    {% endif %}
+  {% endif %}
+  <div id="login-container">
+  {% if wizard.steps.current == 'auth' %}
+    <form method="post" action="">
+      {% csrf_token %}
+      <div class="text-center">
+        <h4>{% trans "Connexion" %}</h4>
+        <br>
+        {{ wizard.management_form }}
+        <p>{% trans "Connectez-vous à l'aide de vos identifiants" %}</p>
+        <p>{% bootstrap_field wizard.form.username show_label=False placeholder="Identifiant" %}</p>
+        <p>{% bootstrap_field wizard.form.password show_label=False placeholder="Mot de passe" %}</p>
+        <button class="btn btn-primary font-weight-bold" type="submit">{% trans "Suivant" %}</button>
+      </div>
+      {# hidden submit button to enable [enter] key #}
+      <input type="submit" value="" class="d-none" />
+    </form>
+    <br>
+    <div>
+      <p><a class="font-weight-bold" href="{% url 'permit_author_create' %}">{% trans "Créer un compte" %}</a></p>
+      <p><a class="font-weight-bold" href="{% url 'password_reset' %}">{% trans "Perdu votre mot de passe ?" %}</a></p>
+    </div>
+  {% elif wizard.steps.current == 'token' %}
+    <p>{% blocktrans trimmed %}Please enter the tokens generated by your token
+      generator.{% endblocktrans %}</p>
+    <form method="post" action="">
+      {% csrf_token %}
+      {% include "two_factor/_wizard_forms.html" %}
+      {% include "two_factor/_wizard_actions.html" %}
+      {% if backup_tokens %}
+        <p class="mt-4">{% trans "As a last resort, you can use a backup token:" %}</p>
+        <p>
+          <button name="wizard_goto_step" type="submit" value="backup"
+                  class="btn btn-secondary btn-block">{% trans "Use Backup Token" %}</button>
+        </p>
+      {% endif %}
+    </form>
+  {% elif wizard.steps.current == 'backup' %}
+    <p>{% blocktrans trimmed %}Use this form for entering backup tokens for logging in.
+      These tokens have been generated for you to print and keep safe. Please
+      enter one of these backup tokens to login to your account.{% endblocktrans %}</p>
+    <form method="post" action="">
+      {% csrf_token %}
+      {% include "two_factor/_wizard_forms.html" %}
+      {% include "two_factor/_wizard_actions.html" %}
+    </form>
+  {% endif %}
+  </div>
+{% endblock %}
